backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
faker==25.0.1
scikit-learn==1.4.2
networkx==3.3
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
